Docs and cleanup for #2997
Auto-merge triggered too soon.
