+++ hapi-fhir-docs/src/main/resources/ca/uhn/hapi/fhir/changelog/5_5_0/2841-word-boundary-text-searches.yaml
@@ -2,4 +2,5 @@
 type: add
 issue: 2841
 title: "The [:text](https://www.hl7.org/fhir/search.html#text) Search Parameter modifier now searches by word boundary of the text content
- as opposed to only searching at the start of the text.  Add * to match word prefixes (e.g. weig* will match weight)."
+   as opposed to only searching at the start of the text when using Lucene/Elasticsearch indexing.
+   Add * to match word prefixes (e.g. weig* will match weight)."
