+++ hapi-fhir-docs/src/main/resources/ca/uhn/hapi/fhir/changelog/5_6_0/3000-Extend-lucene-indexing.yaml
@@ -1,4 +1,5 @@
 ---
 type: add
 issue: 2999
-title: "Lucene/Elasticsearch indexing has been extended to string and token parameters. This can be controlled by the new `setAdvancedLuceneIndexing()` property of DaoConfig."
+title: "Lucene/Elasticsearch indexing has been extended to string, token, and reference parameters.
+  This can be enabled by the new `setAdvancedLuceneIndexing()` property of DaoConfig."
