+++ hapi-fhir-docs/src/main/resources/ca/uhn/hapi/fhir/docs/files.properties
@@ -60,6 +60,7 @@ page.server_jpa.performance=Performance
 page.server_jpa.upgrading=Upgrade Guide
 page.server_jpa.diff=Diff Operation
 page.server_jpa.lastn=LastN Operation
+page.server_jpa.elastic=Lucene/Elasticsearch Indexing
 page.server_jpa.terminology=Terminology
 
 section.server_jpa_mdm.title=JPA Server: MDM
