+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/dao/FulltextSearchSvcImpl.java
@@ -24,6 +24,8 @@ import ca.uhn.fhir.context.FhirContext;
 import ca.uhn.fhir.context.RuntimeSearchParam;
 import ca.uhn.fhir.jpa.api.config.DaoConfig;
 import ca.uhn.fhir.jpa.dao.data.IForcedIdDao;
+import ca.uhn.fhir.jpa.dao.search.AdvancedIndexSearchBuilder;
+import ca.uhn.fhir.jpa.dao.search.HibernateSearchClauseBuilder;
 import ca.uhn.fhir.jpa.model.entity.ResourceLink;
 import ca.uhn.fhir.jpa.model.entity.ResourceTable;
 import ca.uhn.fhir.jpa.model.search.ExtendedLuceneIndexData;
@@ -33,14 +35,9 @@ import ca.uhn.fhir.model.api.IQueryParameterType;
 import ca.uhn.fhir.rest.api.Constants;
 import ca.uhn.fhir.rest.api.server.RequestDetails;
 import ca.uhn.fhir.rest.api.server.storage.ResourcePersistentId;
-import ca.uhn.fhir.rest.param.QuantityParam;
-import ca.uhn.fhir.rest.param.ReferenceParam;
 import ca.uhn.fhir.rest.param.StringParam;
 import ca.uhn.fhir.rest.param.TokenParam;
 import ca.uhn.fhir.rest.server.util.ISearchParamRegistry;
-import com.google.common.collect.Lists;
-import com.google.common.collect.Sets;
-import org.apache.commons.lang3.StringUtils;
 import org.hibernate.search.mapper.orm.Search;
 import org.hibernate.search.mapper.orm.session.SearchSession;
 import org.hl7.fhir.instance.model.api.IAnyResource;
@@ -53,19 +50,15 @@ import org.springframework.transaction.support.TransactionTemplate;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.PersistenceContextType;
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
 
 public class FulltextSearchSvcImpl implements IFulltextSearchSvc {
 	private static final org.slf4j.Logger ourLog = org.slf4j.LoggerFactory.getLogger(FulltextSearchSvcImpl.class);
-	public static final String EMPTY_MODIFIER = "";
 	@Autowired
 	protected IForcedIdDao myForcedIdDao;
 	@PersistenceContext(type = PersistenceContextType.TRANSACTION)
@@ -78,7 +71,7 @@ public class FulltextSearchSvcImpl implements IFulltextSearchSvc {
 	private ISearchParamRegistry mySearchParamRegistry;
 	@Autowired
 	private DaoConfig myDaoConfig;
-
+	private AdvancedIndexSearchBuilder myAdvancedIndexQueryBuilder = new AdvancedIndexSearchBuilder();
 
 	private Boolean ourDisabled;
 
@@ -132,11 +125,6 @@ public class FulltextSearchSvcImpl implements IFulltextSearchSvc {
 		return paramNameToIndexedLink;
 	}
 
-	/**
-	 * These params have complicated semantics, or are best resolved at the JPA layer for now.
-	 */
-	static final Set<String> ourUnsafeSearchParmeters = Sets.newHashSet("_id", "_tag", "_meta");
-
 	@Override
 	public boolean supportsSomeOf(SearchParameterMap myParams) {
 		// keep this in sync with the guts of doSearch
@@ -144,56 +132,12 @@ public class FulltextSearchSvcImpl implements IFulltextSearchSvc {
 
 		requiresHibernateSearchAccess |=
 			myDaoConfig.isAdvancedLuceneIndexing() &&
-				myParams.entrySet().stream()
-					.filter(e -> !ourUnsafeSearchParmeters.contains(e.getKey()))
-					// each and clause may have a different modifier, so split down to the ORs
-					.flatMap(andList -> andList.getValue().stream())
-					.flatMap(Collection::stream)
-					.anyMatch(this::isParamSupported);
+				myAdvancedIndexQueryBuilder.isSupportsSomeOf(myParams);
 
 		return requiresHibernateSearchAccess;
 	}
 
-	private boolean isParamSupported(IQueryParameterType param) {
-		String modifier = StringUtils.defaultString(param.getQueryParameterQualifier(), EMPTY_MODIFIER);
-		if (param instanceof TokenParam) {
-			switch (modifier) {
-				case Constants.PARAMQUALIFIER_TOKEN_TEXT:
-				case "":
-					// we support plain token and token:text
-					return true;
-				default:
-					return false;
-			}
-		} else if (param instanceof StringParam) {
-			switch (modifier) {
-				// we support string:text, string:contains, string:exact, and unmodified string.
-				case Constants.PARAMQUALIFIER_TOKEN_TEXT:
-				case Constants.PARAMQUALIFIER_STRING_EXACT:
-				case Constants.PARAMQUALIFIER_STRING_CONTAINS:
-				case EMPTY_MODIFIER:
-					return true;
-				default:
-					return false;
-			}
-		} else if (param instanceof QuantityParam) {
-			return false;
-		} else if (param instanceof ReferenceParam) {
-			//We cannot search by chain.
-			if (((ReferenceParam) param).getChain() != null) {
-				return false;
-			}
-			switch (modifier) {
-				case	EMPTY_MODIFIER:
-					return true;
-				case Constants.PARAMQUALIFIER_MDM:
-				default:
-					return false;
-			}
-		} else {
-			return false;
-		}
-	}
+
 
 
 	private List<ResourcePersistentId> doSearch(String theResourceType, SearchParameterMap theParams, ResourcePersistentId theReferencingPid) {
@@ -207,7 +151,7 @@ public class FulltextSearchSvcImpl implements IFulltextSearchSvc {
 			)
 			.where(
 				f -> f.bool(b -> {
-					HibernateSearchQueryBuilder builder = new HibernateSearchQueryBuilder(myFhirContext, b, f);
+					HibernateSearchClauseBuilder builder = new HibernateSearchClauseBuilder(myFhirContext, b, f);
 
 					/*
 					 * Handle _content parameter (resource body content)
@@ -232,53 +176,7 @@ public class FulltextSearchSvcImpl implements IFulltextSearchSvc {
 					 * Handle other supported parameters
 					 */
 					if (myDaoConfig.isAdvancedLuceneIndexing()) {
-						// copy the keys to avoid concurrent modification error
-						ArrayList<String> paramNames = Lists.newArrayList(theParams.keySet());
-						for(String nextParam: paramNames) {
-							if (ourUnsafeSearchParmeters.contains(nextParam)) {
-								continue;
-							}
-							RuntimeSearchParam activeParam = mySearchParamRegistry.getActiveSearchParam(theResourceType, nextParam);
-							if (activeParam == null) {
-								// ignore magic params handled in JPA
-								continue;
-							}
-							switch (activeParam.getParamType()) {
-								case TOKEN:
-									List<List<IQueryParameterType>> tokenTextAndOrTerms = theParams.removeByNameAndModifier(nextParam, Constants.PARAMQUALIFIER_TOKEN_TEXT);
-									builder.addStringTextSearch(nextParam, tokenTextAndOrTerms);
-
-									List<List<IQueryParameterType>> tokenUnmodifiedAndOrTerms = theParams.removeByNameUnmodified(nextParam);
-									builder.addTokenUnmodifiedSearch(nextParam, tokenUnmodifiedAndOrTerms);
-
-									break;
-								case STRING:
-									List<List<IQueryParameterType>> stringTextAndOrTerms = theParams.removeByNameAndModifier(nextParam, Constants.PARAMQUALIFIER_TOKEN_TEXT);
-									builder.addStringTextSearch(nextParam, stringTextAndOrTerms);
-
-									List<List<IQueryParameterType>> stringExactAndOrTerms = theParams.removeByNameAndModifier(nextParam, Constants.PARAMQUALIFIER_STRING_EXACT);
-									builder.addStringExactSearch(nextParam, stringExactAndOrTerms);
-
-									List<List<IQueryParameterType>> stringContainsAndOrTerms = theParams.removeByNameAndModifier(nextParam, Constants.PARAMQUALIFIER_STRING_CONTAINS);
-									builder.addStringContainsSearch(nextParam, stringContainsAndOrTerms);
-
-									List<List<IQueryParameterType>> stringAndOrTerms = theParams.removeByNameUnmodified(nextParam);
-									builder.addStringUnmodifiedSearch(nextParam, stringAndOrTerms);
-									break;
-
-								case QUANTITY:
-									break;
-
-								case REFERENCE:
-									List<List<IQueryParameterType>> referenceAndOrTerms = theParams.removeByNameUnmodified(nextParam);
-									builder.addReferenceUnchainedSearch(nextParam, referenceAndOrTerms);
-									break;
-
-								default:
-									// ignore unsupported param types/modifiers.  They will be processed up in SearchBuilder.
-							}
-						}
-
+						myAdvancedIndexQueryBuilder.addAndConsumeAdvancedQueryClauses(builder, theResourceType, theParams, mySearchParamRegistry);
 					}
 					//DROP EARLY HERE IF BOOL IS EMPTY?
